+++ src/plugins/metadataSelector/index.ts
@@ -1,4 +1,2 @@
-export * from './advanced/advancedMetadataSelector.component';
-export * from './advanced/advancedMetadataSelector.interface';
 export * from './no/noMetadataSelector.component';
 export * from './no/noMetadataSelector.interface';